Müsteqil çalışma n1 (alternatif a // b komutu)

+++ yuvarlama.py
@@ -1,4 +1,4 @@
-x = int(input("x:"))
+x = float(input("x:"))
 y = float(input("y:"))
 z = x / y
 # qalıq mod işlemi
